update API delete user

+++ server/api/controllers/user/controller.js
@@ -51,15 +51,22 @@ export class Controller {
 
     deleteUserByID(req,res){
         const {id} = req.params
-        userModel.findByIdAndDelete(id)
-        .then((userDeleted)=>{
-            if(userDeleted){
-                res.status(200).send("user deleted")
-            }
-            else{
-                res.status(404).send("User not found")
+        postModel.deleteMany({user : id}).then(deletedPost=>{
+            if(deletedPost){
+                userModel.findByIdAndDelete(id)
+                .then((userDeleted)=>{
+                    if(userDeleted){
+                        res.status(200).send("user deleted")
+                    }
+                    else{
+                        res.status(404).send("User not found")
+                    }
+                })
+            }else{
+                res.status(404).send("Cant delete post so cant delete user")
             }
         })
+
     }
 
     getUserByID(req,res){
